requirements.txt
discord.py==2.3.2
cerberus==1.3.4
dateparser
psycopg2